moved all pages to the page folder
