--- src/components/NotFound.tsx
@@ -1,11 +1,0 @@
-export const NotFound = () => (
-  <div className="bg-[url('./assets/404.jpg')] bg-contain	bg-no-repeat bg-center h-screen relative -z-10">
-    <a
-      href="https://www.vecteezy.com/free-vector/404-page"
-      rel="noreferrer noopener"
-      className="absolute end-0	bottom-0 text-sm"
-    >
-      404 Page Vectors by Vecteezy
-    </a>
-  </div>
-);
