+++ src/components/RequestArea.tsx
@@ -1,0 +1,5 @@
+const RequestArea = () => {
+  return <div></div>;
+};
+
+export default RequestArea;
