+++ src/pages/WelcomePage/Cover.tsx
